requirements.txt
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu
diffusers
transformers
accelerate
safetensors
streamlit
pillow
